Move istio util duncs to pkg/util/istio
